refactor: Created separate files for each module, and renamed ComponentHooks to Controllers
